Remove most use of @-boxes + small api changes.
Fix #5.

Most @-boxes have been replaced by RCMut or RWArc depending of the situation.
The only @ boxes still existing are the one needed so that the user can keep a usefully typed
reference to something which is converted to a trait-object by the engine (cameras and
post-processing effects).

Here are a few additional api changes:
    - user-defined cameras are now possible!
    - SPACE does not switch to wireframe mode any more.
    - TAB does not switch between arc-ball and first-person any more.
    - method call cannot be chained on Object any more.
    - `set_loop_callback` becomes `render_loop` which must be called to start the (infinite) render
      loop.

This adds a demo to show how to switch to wireframe mode.