Breaking changes: update to the last rust-nightly.
Version of rustc: 0.12.0-pre-nightly (4d4eb1023 2014-08-02 23:36:09 +0000).

Until now, the iterator-based rendering took advantage of a
compiler bug (it did not check that the iterator implemented the
`Iterator` trait).

Now that the compiler bug is fixed, it is no longer possible to
use this method.

Instead, the user must now use a plain old `while` loop until
the `render` method returns false (or the `should_close` method
returns true). This is actually easier than the iterator thingy
since it does not have to borrow the main window (which is now
usable on the loop).

Also, the examples have their own `Cargo.toml` on the `examples`
directory.