webgl_canvas: Put key and wheel event listeners on canvas element
This causes key and wheel events to be sent to the Kiss3d context only
if the canvas element has focus, to avoid interfering with other HTML
elements.

The code now also adds a `tabindex` attribute to the canvas in order to
allow keyboard events to be received, due to the keyboard events now
being added to the canvas element instead of the window.