camera: improved arcball mouse zoom
This commit improves the mouse zoom for the arcball camera. Currently
the zoom is toward the center point, regardless of the mouse position.
With this change, the zoom is now focused toward the part of the scene
under the cursor.

We achieve this by temporarily translating the camera to change the
focus point.