Add line width to planar line renderer
The planar line renderer does not currently set the line width; when
rendering, it inherits any line width that was set by the 3D line
renderer (if it happened to be used).

This commit adds a line width setting to the planar line renderer; the
`window::set_line_width` method sets it together with the 3D line
width.