Always enable optimizations on image loader
The examples that involve textures show a noticeable delay at start
in dev mode; the delay is not there in release mode. I tracked this
down to the image library. This commit changes the profile for the
image library and the jpg and png decoding libraries to use
optimizations and disable debug assertions.