Remove `RenderLoopClosure` trait
I got better at Rust, and realized that we don't need this trait (which
I added a year ago). We can just make `RenderLoopClosureImpl` implement
`FnMut` and keep the method signatures the same as before.

But then it turns out that you can't implement the `Fn` without some
unstable features, so I came up with something else.

This reverses a lot of the changes from #286, but still has the same
effect: `cargo run --example persistent_point_cloud` does not give an
error when the window is closed (which is what drove me to make that PR
in the first place.)