Fix for github.com/sebcrozet/kiss3d/issues/98
The normalized OpenGL clip coordinates are defined as follows. The origin
is at the center of the view. The horizontal or x-axis ranges left to right
from -1 to 1. The vertical or y-axis also ranges bottom to top from -1 to
1.

Device or screen coordinates are a rasterized version of the following. The
origin is the top left corner of the view port (window). For a view port of
width ϝ and height η, the horizontal or x-axis ranges left to right from 0
to ϝ, and the vertical axis ranges top to botton from from 0 to η.

Clip coordinates map to device coordinates in the following way. Let
(x, y) be a position in clip coordinates, and (χ, υ) be the same position
in device coordinates. This means that χ = [(ϝ - 0)/(1 - -1)](x - 1) and
υ = [(0 - η)/(1 - -1)](y - 1). These reduce to χ = (ϝ/2)(x - 1) and
υ = -(η/2)(y - 1), respectively.

The inverse of the transformations are the following. x = 1 + (2/ϝ)x and
y = 1 - (2/η)υ.

In both the horizontal and vertical directions, the scale factor is twice
(2x) what the previous version in the function method
text::renderer::TextRender::render.

This should be sufficient to close issue
https://github.com/sebcrozet/kiss3d/issues/98.